system-monitor: Don't read argv beyond the boundary
If there are no elements in argv, the first item could be itself null.
And thus reading the next item is illegal. So break the loop once null is found.

Also as both elements of secure_arguments are required to have some value
insert a blank in case null is reached early.

https://bugzilla.gnome.org/show_bug.cgi?id=782487