Add progress indicator when the storage panel is loading
This commit changes the spinner into a progress bar.  It is updated with
information from the storage analyzer.

The storage analyzer was changed to add another message type, updating
the size.  This is emitted as quickly as possible, so that the user gets
the most accurate and optimistic view of the progress.