build: Improve compiled resources
The list of files used by the resources file has been added, so
their existence is checked during the resource generation.

The gresources vala parameter has been also removed, because it
has the same effect as it not necessary anymore.

The meson's `gnome` module has also been grouped with the i18n
module just for consistency.