build: Improve vapi files handling
The common directory for holding vapi files is usually the `vapi`
directory in the root source directory.

The `vapis` directory have been moved to the root source directory
from the `src` directory, and has been renamed to `vapi`.

meson is also able to detect vala dependencies, so that feature is
used instead of vala arguments.