build: Use meson's conventions
The convention when naming dependencies is to use the '_dep' suffix,
which helps detecting dependency variables. The variables have been
renamed to follow this convention. The dependencies that are also
used once, are declared in the target's dependency array which helps
meson processing less and also one less variable is needed.

There is also a convention when naming libraries which is by using
the `lib` prefix. The libraries have renamed accordingly.

Finally, usually a 2 space indentation is used. This has also been
changed.