graph-stack-switcher: Better change active button on scrolling
With this commit GraphStackSwitcher will switch active button
when subview will occupy more than half of displayed window.

This too solve problem when is last subview smaller then half of
window but window is scrolled to the end using bottom tolerance.

This is much more intuitive for users and gives it a sense of user
experience.