system-monitor: maintain a stable map of processes
Maintain a pid → Process map. On each update compare the old and
the new list of pids. Add news Process objects for new pids and
remove the Process objects from the map for pids that got removed.
For matching pids, make sure that also the start time matches, so
we can detect (unlikely, but possible) roll-over, i.e. a process
died and a new process got assigned the very same pid.
This should help to reduce computations, i.e. the full command
line, over and over. Also we can, in the future, detect if a
process is containerized and memorize that status in the Process
object.