process: support to lookup app_id for flatpaks
Add an app_id property, that on the first time it is access, will
try to open the .flatpak-info keyfile, via the /proc/<pid>/root
link. If this succeeds read the keyfile and return the "name" of
the "Application" section. For now manually add a .desktop suffix
so that this hopefully matches the AppInfo.get_id() of the
corresponding desktop file.