introduce empty graph view
To be able to remove the dependency on libdazzle, Usage
requires a new graph view. That, however, will be
implemented using GTK4's new render nodes.

In order to port Usage to GTK4, a fake graph view is
introduced here, and everything of the old code that is
not needed anymore, is removed.

closes #40, #53 as obsolete

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>