replace loading notifications with progress bar
Currently, the notification bar for the Storage view
is merely used for conveying the loading state of the
initial directory list.

While it and its custom widgetry is best replaced with
something more aligned with the GNOME HIG, neither Toasts
nor Banners are well suited for this task.

Therefore, it could be replaced with a pulsing progress
bar at the top of the Storage view that, while at it,
could also indicate the loading state of every other
directory scan.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>