hide empty row for available space in storage list
A fake storage item is added to the list model to show
the available space in the graph, but for the list itself
file row is not supposed to exist and currently an empty
row.

That, however, leads to it being able to have margins,
and to be hovered and to be selected. As there has to be
a row for this entry, a workaround is to at least make it
invisible.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>