remove double tracking of current directory in storage panel
As the stack list already keeps track of the directory
list, there is no need to also keep a queue of which is
the current directory in the storage view itself.

And as the directory stack always contains every entry
between $HOME and the current directory, the refresh
depth is obsolete.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>