storage: stop referencing variables from a lambda's parent scope
If a variable is referenced from outside a lambda
expression, it is passed along with a struct pointer that
could lead to reference cycles and double counting.

Even though, it does not happen here, there is no
guarantee it won't in the future, and in fact it will
with the GTK4 port.

In this case, the variable is also the sender of the
signal. Therefore, the sender can be used instead.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>