grant file system access to scanned user directories
The storage view expects the user directories that are to
be scanned to exist in the directory structures it
perceives, and thus crashes within the bubblewrap sandbox.

And as the file system access is also needed for the
feature to delete from within the storage view, it is
best to just be granted.

fixes #104

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>