add vertical bounds to graph lines
The graph lines can have a vertical expansion far above
the stroke width with a high angle. Then, they do not
necessarily align with the next section's line. This is
especially noticeable if the next angle is closer to 0.

Caping the bounds of the lines to values close to the
stroke width of the lines, reduces the visual overshoot.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>