performance-view: normalize CPU load to 100 %
With n threads the load can get up to n. Thus, deviding
by that amount leads to relative numbers with 1 being
at maximum load.

Decreasing the displayed load values introduces the need
to increase the precision of the formatter in performance
list.

To accomodate for the now lower load values, the minimum
load to be displayed is updated as well.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>