app-item: introduce API for checking killability
The AppItem itself is not yet aware of whether it is
killable or not, but it is cleaner to do so, than having
those checks in the UI parts.

Furthermore, this moves this check to a single point,
making it easier to improve the logic in the future and
making it saver against forgotten checks.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>