move performance sub views to individual views
Additionally, this renders multiple custom switcher
widget unneeded, as the will be replaced by a generic
ListView in a future sidebar, as well as the SubView and
AnimatedScrolledWindow, because each view will stay at
toplevel.

Without the custom switcher widgetry, the window can
shrink in height, so the Views' and Window's height-
requests are adjusted.

Closes: https://gitlab.gnome.org/GNOME/gnome-usage/-/issues/84 obsolete
Closes: https://gitlab.gnome.org/GNOME/gnome-usage/-/issues/85 obsolete
Closes: https://gitlab.gnome.org/GNOME/gnome-usage/-/issues/86 obsolete

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>