display swap graph only if swap available
If no swap is available, currently, the memory graph does
not display swap. However, that is due to a bug, as it is
not calculated for no RAM available, instead of no swap.

By only adding new points to this graph if swap does
exist, the current behaviour is retained, while glitches
due to invalid values are stopped from appearing.

Closes: https://gitlab.gnome.org/GNOME/gnome-usage/-/issues/120

Backported 5a275eb996cab38a4d1894268bb0bec86252bd23

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>