add app dialog for process rows
This adds another layer between process rows and the quit
dialogues, as clicking a row could be also an intend to
know more about the process and app in question.

It also helps with providing a better interface for when
SIGTERM is not enough.

This dialogue should be extended information-wise in the
future.

Closes: https://gitlab.gnome.org/GNOME/gnome-usage/-/issues/20

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>