app-item: prioritise app infos without nodisplay
The assumption is that the main entry to the app is
supposed to be the preferred way to display it, as it is
the way the app is shown in the app grid, and generally
has the icon and name associated with the app and not a
single aspect of it.

An example would be Papers vs. its print preview.

Backported 3c50ea308e81d37234fc337eff54a1a707aa0000

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>