process: simplify cgroup reading
The cgroup was read whenever a process was processed for
data update. However, the cgroup normally does not change
that often. So, loading only once should work in most
cases.

It also was unnecessary to create a fallback app ID, if
the app ID was successfully retrieved. And
AppInfo.app_info_for_process was not null-safe.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>