performance: change default minimum load to 0
The app-minimum-load often confused users as to why apps
did not show up.

The weirdness is most apparent when you start searching
and with the filter applied more apps are listed than
with the filter.

A minimum load of 0 means any app greater than no load is
considered running. It still excludes app not using any
CPU cycles.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>