system-monitor: do active use counting and track screensaver state
There is no need to update everything at full speed, when
no window is active. An active window is one that is
visible, i.e. if none is active, it means they are
hidden.

It makes sense to slow down the data update, but keep
collecting, because some views, such as the graphs depend
on historic data and not just the current one. It is a
common pattern to look at a system monitoring app, such as
Usage, shorty after you notice an issue, e.g. fans turning
on. Not aggregating this data at all would make Usage
considerably less useful.

When the screen is locked, however, the user is even less
likely to need to the graph. As such, the system monitor
can be turned off with important power draining
improvements for users having Usage running in background.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>