use power saving mode for disabling system monitor
While it is useful to have the graph data being available
when looking at Usage after noticing an issue, most data
aggregated is discarded with only the last 45 s being
used for the graphs.

It is unreasonable to throw power at this potentially
useful, but most likely never used, data collection,
when the user or their system explicitly requested power
saving.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>