update development app ID to org.gnome.Usage.Devel
The app ID until now was org.gnome.UsageDevel, which
does not follow the style by others. Due to this app
being not very usable in a sandbox and this one being
mostly for targeted testing, and it not being available
after the gnome-nightly purge for a while now, there is
probably no drawback to this update.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>